test3

+++ 16_subseq/maxSeq.c
@@ -5,10 +5,10 @@ size_t maxSeq(int * array, size_t n) {
   if (n<0) return EXIT_FAILURE;
   if (array==NULL) return EXIT_FAILURE;
   if (n==0) return 0;
-  int maxCount1=1;
-  int maxCount2=1;
-  int temp;
-  for (int i=0; i<n; i++){
+  size_t maxCount1=1;
+  size_t maxCount2=1;
+  size_t temp;
+  for (size_t i=0; i<n; i++){
     if (array[i]<array[i+1]) {
       maxCount1++;
 	  } else {
